fix directory layout
