+++ models/user.go
@@ -1,0 +1,26 @@
+package models
+
+type UserLoginResponse struct {
+	Response
+	UserID int64  `json:"user_id,omitempty"`
+	Token  string `json:"token,omitempty"`
+}
+
+type UserInfo struct {
+	ID              int64  `json:"id"`
+	Name            string `json:"name"`
+	FollowCount     int64  `json:"follow_count"`
+	FollowerCount   int64  `json:"follower_count"`
+	IsFollow        bool   `json:"is_follow"`
+	Avatar          string `json:"avatar"`
+	BackgroundImage string `json:"background_image"`
+	Signature       string `json:"signature"`
+	TotalFavorited  int64  `json:"total_favorited"`
+	WorkCount       int64  `json:"work_count"`
+	FavoriteCount   int64  `json:"favorite_count"`
+}
+
+type UserInfoResponse struct {
+	Response
+	User *UserInfo `json:"user,omitempty"`
+}
